add gen0

+++ include/gc/parameters.h
@@ -7,7 +7,7 @@
 #define MAX_ALLOC_SIZE ((size_t)GIGABYTE)
 #endif
 
-#define GEN0_SPACE_SIZE (MAX_ALLOC_SIZE / 3)
+#define GEN0_SPACE_SIZE (((size_t)MAX_ALLOC_SIZE) / 3)
 #define GEN1_SPACE_SIZE (GEN0_SPACE_SIZE * 2)
 
 #endif // PARAMETERS_H
